requirements.txt
dataclasses
datasets
matplotlib
neptune-client
numpy
pandas
pillow
pyyaml
sympy
torch
torchvision
tqdm